+++ engine/src/test/java/org/teiid/query/processor/TestProcessor.java
@@ -230,7 +230,7 @@ public class TestProcessor {
         }
     }
         
-    public static int doProcess(ProcessorPlan plan, ProcessorDataManager dataManager, List[] expectedResults, CommandContext context) throws Exception {
+    public static long doProcess(ProcessorPlan plan, ProcessorDataManager dataManager, List[] expectedResults, CommandContext context) throws Exception {
     	BufferManager bufferMgr = context.getBufferManager();
     	if (bufferMgr == null) {
 	        BufferManagerImpl bm = BufferManagerFactory.createBufferManager();
@@ -255,7 +255,7 @@ public class TestProcessor {
         	context.setQueryProcessorFactory(new QueryProcessorFactoryImpl(bufferMgr, dataManager, new DefaultCapabilitiesFinder(), null, context.getMetadata()));
         }
         TupleBuffer id = null;
-        int rowCount = 0;
+        long rowCount = 0;
         try {
             QueryProcessor processor = new QueryProcessor(plan, context, bufferMgr, dataManager);
             //processor.setNonBlocking(true);
@@ -304,7 +304,7 @@ public class TestProcessor {
         
         // Create QueryResults from TupleSource
         TupleSource ts = tsID.createIndexedTupleSource();
-        int count = tsID.getRowCount();   
+        long count = tsID.getRowCount();   
 
         // Compare actual to expected row count
         assertEquals("Did not get expected row count: ", expectedResults.length, count); //$NON-NLS-1$
